# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Library Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  Admin: 'Admin',
  Student: 'Student',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  Admin: 'Admin',
  Student: 'Student',
} as const;

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Book {
  id: number;
  title: string;
  author: string;
  isbn: string;
  category: string;
  quantity: number;
  availableQuantity: number;
  createdAt: string;
}

export interface BookInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  author: string;
  /** @minLength 1 */
  isbn: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 1 */
  quantity: number;
}

export interface BookUpdate {
  title?: string;
  author?: string;
  isbn?: string;
  category?: string;
  /** @minimum 0 */
  quantity?: number;
}

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  active: 'active',
  returned: 'returned',
  overdue: 'overdue',
} as const;

export interface Transaction {
  id: number;
  userId: number;
  bookId: number;
  issueDate: string;
  returnDate: string;
  /** @nullable */
  actualReturnDate?: string | null;
  /** @nullable */
  fineAmount?: number | null;
  status: TransactionStatus;
  book: Book;
  user: User;
}

export interface IssueInput {
  bookId: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  dueDays?: number;
}

export interface Notification {
  id: number;
  userId: number;
  message: string;
  isRead: boolean;
  timestamp: string;
}

export type AdminDashboardBooksByCategoryItem = {
  category: string;
  count: number;
};

export interface AdminDashboard {
  totalBooks: number;
  totalUsers: number;
  activeRentals: number;
  overdueRentals: number;
  totalFinesCollected: number;
  pendingFines: number;
  booksByCategory: AdminDashboardBooksByCategoryItem[];
  recentTransactions: Transaction[];
}

export interface StudentDashboard {
  currentlyBorrowed: number;
  totalBorrowed: number;
  outstandingFines: number;
  overdueBooks: number;
  activeTransactions: Transaction[];
  unreadNotifications: number;
}

export type ListBooksParams = {
search?: string;
category?: string;
};