# lib/api-zod/src/generated/types/studentDashboard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Library Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from './transaction';

export interface StudentDashboard {
  currentlyBorrowed: number;
  totalBorrowed: number;
  outstandingFines: number;
  overdueBooks: number;
  activeTransactions: Transaction[];
  unreadNotifications: number;
}